+++ docs/Release Notes/Change Log.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [Unreleased]
+
+#### Added
+
+- [#20](https://github.com/mabel-dev/opteryx/issues/20) Split query planner and query plan into different modules. ([@joocer](https://github.com/joocer]))
+- [#165](https://github.com/mabel-dev/opteryx/issues/165) Support S3/MinIO data stores for blobs. ([@joocer](https://github.com/joocer]))
+
 ### [0.0.2] - 2022-06-03
 
 #### Added